requirements.txt
apache-airflow-providers-apache-kafka
kafka-python
pandas
requests
sqlalchemy